https://applitools.atlassian.net/wiki/spaces/Java/pages/1540555/Images+SDKs+-+JavaScript?showComments=true&showCommentArea=true
(Optional) Setting app environment attributes
Normally applitools eyes SDKs automatically identifies the environment that is used to run the tests and creates a separate baseline for each environment. In order to override app environment parameters and ‘force’ Eyes to compare results of different environments, you can explicitly call ‘setAppEnvironment’:

hostOs - The OS name and its version or any other os-alike identifier. hostApp - The application name, if there is, such as browser name or any other app-alike identifier.

This functionality can be used to ‘force’ eyes to preset the hostOS and hostApplication so identical tests on different environments will be unified under the same baseline. Note that using this in such way will most likely cause the tests to fail, since in most cases there are some differences between how the AUT is presented in the different environments.